Move stitching from Replicated->Unstructured Mesh
Yes, right now we have to serialize to do this.  But at least we can
then re-distribute later; better for stitching DistributedMesh to be
inefficient rather than impossible.
